requirements.txt
aiohttp>=3.8
websockets>=10.0
pandas>=1.5
rich>=13.0
python-dotenv>=0.21
numpy>=1.24
python-telegram-bot>=20.0
Pillow>=9.0
certifi
